+++ packages/adapter-interfaces/src/types/Client.ts
@@ -1,6 +1,7 @@
 import { z } from "@hono/zod-openapi";
 import { connectionSchema } from "./Connection";
 import { applicationSchema } from "./Application";
+import { tenantSchema } from "./Tenant";
 
 const ClientDomainSchema = z.object({
   domain: z.string(),
@@ -12,29 +13,11 @@ const ClientDomainSchema = z.object({
     .optional(),
 });
 
-const BaseClientSchema = z.object({
+const ClientSchema = z.object({
   ...applicationSchema.shape,
   domains: z.array(ClientDomainSchema),
-  tenant: z.object({
-    name: z.string(),
-    audience: z.string().optional(),
-    logo: z.string().optional(),
-    primary_color: z.string().optional(),
-    secondary_color: z.string().optional(),
-    sender_email: z.string(),
-    sender_name: z.string(),
-    support_url: z.string().optional(),
-    language: z.string().length(2).optional(),
-  }),
-});
-
-export const ClientSchema = BaseClientSchema.extend({
+  tenant: tenantSchema,
   connections: z.array(connectionSchema),
 });
 
-export const PartialClientSchema = BaseClientSchema.extend({
-  connections: z.array(connectionSchema.partial()),
-});
-
 export type Client = z.infer<typeof ClientSchema>;
-export type PartialClient = z.infer<typeof PartialClientSchema>;