+++ packages/adapter-interfaces/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @authhero/adapter-interfaces
 
+## 0.11.3
+
+### Patch Changes
+
+- Simplify client
+
 ## 0.11.2
 
 ### Patch Changes